Refaactor main networking code
Refactor code to improve readability and modular
